Change type of indexing variables to uint32_t in multiple files
The type of several indexing variables used in for-loop iterations is changed from size_t to uint32_t. This change was made in various source files, including but not limited to SaveGenJob.cpp, SCBE_X64.cpp, Parser_Declaration.cpp, etc. The commit ensures uniform type for all indexing variables enhancing code consistency.

+++ src/Ast.cpp
@@ -46,7 +46,7 @@ Utf8 Ast::enumToString(TypeInfo* typeInfo, const Utf8& text, const Register& reg
     if (typeEnum->declNode->hasAttribute(ATTRIBUTE_ENUM_FLAGS))
     {
         SWAG_ASSERT(typeEnum->rawType->isNative());
-        for (size_t i = 0; i < typeEnum->values.size(); i++)
+        for (uint32_t i = 0; i < typeEnum->values.size(); i++)
         {
             const auto value = typeEnum->values[i];
             bool       ok    = false;
@@ -106,7 +106,7 @@ Utf8 Ast::enumToString(TypeInfo* typeInfo, const Utf8& text, const Register& reg
         if (scoped)
             result = typeInfo->name;
         result += ".";
-        for (size_t i = 0; i < typeEnum->values.size(); i++)
+        for (uint32_t i = 0; i < typeEnum->values.size(); i++)
         {
             const auto value = typeEnum->values[i];
             SWAG_ASSERT(value->value);
