+++ src/OsWin32.cpp
@@ -348,7 +348,7 @@ namespace OS
 
         // Remove unwanted characters
         char* pz = messageBuffer;
-        for (size_t i = 0; i < size; i++)
+        for (uint32_t i = 0; i < size; i++)
         {
             if (pz[i] < 32)
                 pz[i] = 32;
